client_voice_apps/voice-synthesis/requirements.txt
kafka-python==2.0.1
pydub==0.24.0